LayoutTests/fast/js/delete-multiple-global-blocks.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
<script src="resources/js-test-pre.js"></script>
</head>
<body>
<script>
description(
"Tests whether delete works properly across multiple global blocks in the presence of redefined variables."
);

a = 1;
</script>
<script>
var a;
delete a;

function test()
{
    return this.a ? true : false;
}

shouldBeFalse("test()");
</script>
<script src="resources/js-test-post.js"></script>
</body>
</html>
